Refactor OpenAI utilities and enhance PromptTemplate
- Move OpenAI-specific utilities to dedicated module
- Add MD5 hash properties to PromptTemplate for message fingerprinting
- Update imports across affected modules and tests
- Remove redundant prompt utility module
