+++ languru/utils/common.py
@@ -18,7 +18,6 @@ from typing import (
     cast,
 )
 
-from openai.types.chat import ChatCompletionMessageParam
 from pydantic import BaseModel
 from rich import box
 from rich import print as rich_print
@@ -30,7 +29,7 @@ from languru.config import console, logger
 from languru.types.chat.completions import Message
 
 if TYPE_CHECKING:
-    from openai.types.chat.chat_completion import ChatCompletion
+    from openai.types.chat import ChatCompletionMessageParam
 
 T = TypeVar("T")
 
@@ -117,15 +116,6 @@ def remove_punctuation(input_string: Text, extra_punctuation: Text = "") -> Text
     return input_string.translate(translator)
 
 
-def ensure_openai_chat_completion_content(chat_completion: "ChatCompletion") -> Text:
-    """Ensure that the chat completion response content is returned."""
-
-    chat_answer = chat_completion.choices[0].message.content
-    if chat_answer is None:
-        raise ValueError("Failed to get a response content from the OpenAI API.")
-    return chat_answer
-
-
 def ensure_list(value: Any) -> List:
     if isinstance(value, Sequence):
         return list(value)
@@ -134,38 +124,11 @@ def ensure_list(value: Any) -> List:
     return [value]
 
 
-def ensure_openai_chat_completion_message_params(
-    messages: Union[
-        Sequence["Message"],
-        Sequence[Dict[Text, Any]],
-        Sequence[ChatCompletionMessageParam],
-    ]
-) -> List[ChatCompletionMessageParam]:
-    """Ensure that the chat completion messages are in the correct format."""
-
-    _messages: List[ChatCompletionMessageParam] = []
-    for m in messages:
-        if isinstance(m, Message):
-            m_dict = m.model_dump()
-            if m_dict.get("role") not in ["user", "assistant", "system"]:
-                raise ValueError(f"Invalid role: {m_dict.get('role')}")
-            _messages.append(m_dict)  # type: ignore
-        elif isinstance(m, Dict):
-            if m.get("role") not in ["user", "assistant", "system"]:
-                raise ValueError(f"Invalid role: {m.get('role')}")
-            _messages.append(m)  # type: ignore
-        elif isinstance(m, ChatCompletionMessageParam.__args__):
-            _messages.append(m)
-        else:
-            raise ValueError(f"Invalid message type: {m}")
-    return _messages
-
-
 def display_messages(
     messages: Union[
         Sequence["Message"],
         Sequence[Dict[Text, Any]],
-        Sequence[ChatCompletionMessageParam],
+        Sequence["ChatCompletionMessageParam"],
     ],
     *,
     is_print: bool = True,
