+++ languru/utils/openai_utils.py
@@ -1,7 +1,63 @@
-from typing import Text
+from typing import Any, Dict, List, Sequence, Text, Union
 
+from openai.types.chat import ChatCompletionMessageParam
+from openai.types.chat.chat_completion import ChatCompletion
 from pyassorted.string.rand import rand_str
+
+from languru.types.chat.completions import Message
 
 
 def rand_chat_completion_id() -> Text:
     return f"chatcmpl-{rand_str(29)}"
+
+
+def ensure_chat_completion_message_params(
+    messages: (
+        Sequence[ChatCompletionMessageParam]
+        | Sequence[Dict[Text, Any]]
+        | ChatCompletionMessageParam
+        | Dict[Text, Any]
+    )
+) -> List[ChatCompletionMessageParam]:
+    """Ensure the messages parameter is a list of ChatCompletionMessageParam objects."""
+
+    if not messages:
+        raise ValueError("The messages parameter is required.")
+
+    return list([messages] if isinstance(messages, Dict) else messages)  # type: ignore
+
+
+def ensure_openai_chat_completion_message_params(
+    messages: Union[
+        Sequence["Message"],
+        Sequence[Dict[Text, Any]],
+        Sequence[ChatCompletionMessageParam],
+    ]
+) -> List[ChatCompletionMessageParam]:
+    """Ensure that the chat completion messages are in the correct format."""
+
+    _messages: List[ChatCompletionMessageParam] = []
+    for m in messages:
+        if isinstance(m, Message):
+            m_dict = m.model_dump()
+            if m_dict.get("role") not in ["user", "assistant", "system"]:
+                raise ValueError(f"Invalid role: {m_dict.get('role')}")
+            _messages.append(m_dict)  # type: ignore
+        elif isinstance(m, Dict):
+            if m.get("role") not in ["user", "assistant", "system"]:
+                raise ValueError(f"Invalid role: {m.get('role')}")
+            _messages.append(m)  # type: ignore
+        elif isinstance(m, ChatCompletionMessageParam.__args__):
+            _messages.append(m)
+        else:
+            raise ValueError(f"Invalid message type: {m}")
+    return _messages
+
+
+def ensure_openai_chat_completion_content(chat_completion: "ChatCompletion") -> Text:
+    """Ensure that the chat completion response content is returned."""
+
+    chat_answer = chat_completion.choices[0].message.content
+    if chat_answer is None:
+        raise ValueError("Failed to get a response content from the OpenAI API.")
+    return chat_answer
